Car on the track + kerbs!
Thus far the car is just being "placed", so it's not really "moving" in the world. It is set up just so that it can be seen. The kerbs are bitmaps, but there is a certain amount of "jerkiness" - it may be that I have to go back to my former method.