Fixed problems with running on phone
Actually, I had not updated the Pebble App on the phone and that was causing the game not to run. The game now runs, but with minimal UI and you can run only one race. Next is to change the boost UI slightly - at the moment it is hard to see it against the barrier pattern.