AI that can be controlled individually
This is done with carPtr->boostRank and carPtr->boostChance. At the moment it is set so that the behaviour is the same as before. Eventually I will find a way to set these based on the difficulty level. In general a high boostRank and low boostChance will make the game easier.