Fixed a bug in AI steering.
This bug allowed the AI cars to travel on the right-hand barrier. This commit also includes a slight update to the positioning of the car (basically moving it further back on the screen so the player can see more of what is going on ahead). Finally, a very slight change to the indicator that shows which car is the player car.